Create 02_17-array4.c
in this case, note that the function returns a triplet vector.  This is done by passing the result vector back through the function call.  Techanically, we're passing a "pointer" to the function, the function messes with the data, and then when the function finishes, the result is sitting in the memory we "pointed to".  Pointers are tricky though - there's much more to the discussion than this example.